@@ -1,16 +1,15 @@
 package com.delzor.zb.csakbusz.adapters
 
-import android.support.v7.widget.RecyclerView
+import android.content.Intent
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.ViewGroup
+import android.widget.ImageView
+import android.widget.TextView
 import com.delzor.zb.csakbusz.*
 import com.delzor.zb.csakbusz.activities.BusStopsLocationMapActivity
 import com.delzor.zb.csakbusz.activities.LineDataActivity
 import com.delzor.zb.csakbusz.activities.StopDataActivity
-import kotlinx.android.synthetic.main.list_item_busstop.view.*
-import org.jetbrains.anko.imageResource
-import org.jetbrains.anko.sdk25.coroutines.onClick
-import org.jetbrains.anko.startActivity
 
 class BusStopLineListAdapter(var stopList: MutableList<BusStop>? = null, var  lineList: MutableList<SimpleLine>? = null) : RecyclerView.Adapter<CustomViewHolder>() {
     val lines = lineList != null
@@ -28,32 +27,35 @@ class BusStopLineListAdapter(var stopList: MutableList<BusStop>? = null, var  li
         val cnt = holder.view.context
         if(lines){
             val line = lineList!![position]
-            holder.view.tvBusStopName.text = "${line.num} ${line.name.substring(2).replace("-"," - ")}"
-            holder.view.ivStopOrBus.imageResource = R.drawable.bus_marker
+            holder.view.findViewById<TextView>(R.id.tvBusStopName).text = "${line.num} ${line.name.substring(2).replace("-"," - ")}"
+            holder.view.findViewById<ImageView>(R.id.ivStopOrBus).setImageResource(R.drawable.bus_marker)
 
-            holder.view.onClick{
+            holder.view.setOnClickListener{
                 Data._testLineNum = line.num
-                cnt.startActivity<LineDataActivity>(
-                        "TITLE" to line.name
-                )
+                Intent(cnt, LineDataActivity::class.java).apply {
+                    putExtra("TITLE", line.name)
+                }.also { cnt.startActivity(it) }
             }
 
 
         }else{
             val busStop = stopList!![position]
-            holder.view.tvBusStopName.text = busStop.name
+            holder.view.findViewById<TextView>(R.id.tvBusStopName).text = busStop.name
 
-            holder.view.onClick {
+            holder.view.setOnClickListener {
                 Data.fetchStopSpot(busStop.id) {
                     //println("SIZE:::: "+Data.currStopSpots.size)
                     if (Data.currStopSpots.size == 1) {
                         // ha csak 1 buszmeg van kihagyja a buszválasztó mapot
-                        cnt.startActivity<StopDataActivity>("STOPSPOT_ID" to Data.currStopSpots[0].id)
+                        Intent(cnt, StopDataActivity::class.java).apply{
+                            putExtra("STOPSPOT_ID", Data.currStopSpots[0].id)
+                        }.also { cnt.startActivity(it) }
+
                         //println("CSAK EGY")
                     } else {
-                        cnt.startActivity<BusStopsLocationMapActivity>(
-                                "BUS_STOP_ID" to busStop.id
-                        )
+                        Intent(cnt, BusStopsLocationMapActivity::class.java).apply{
+                            putExtra("BUS_STOP_ID", busStop.id)
+                        }.also { cnt.startActivity(it) }
                         //println("TÖBB")
                     }
                     Data._testStopSpotName = busStop.name
