@@ -1,13 +1,12 @@
 package com.delzor.zb.csakbusz.adapters
 
-import android.support.v7.widget.RecyclerView
+import android.content.Intent
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.ViewGroup
+import android.widget.Button
 import com.delzor.zb.csakbusz.activities.BusViewMapActivity
 import com.delzor.zb.csakbusz.R
-import kotlinx.android.synthetic.main.list_item_button.view.*
-import org.jetbrains.anko.sdk25.coroutines.onClick
-import org.jetbrains.anko.startActivity
 
 class BuslistAdapter(var itemList: MutableList<String>) : RecyclerView.Adapter<CustomViewHolder>() {
     override fun getItemCount(): Int {
@@ -23,15 +22,12 @@ class BuslistAdapter(var itemList: MutableList<String>) : RecyclerView.Adapter<C
     override fun onBindViewHolder(holder: CustomViewHolder, position: Int) {
         val cnt = holder.view.context
         val btnTxt = itemList[position]
-        holder.view.btnListItem.text = btnTxt
+        holder.view.findViewById<Button>(R.id.btnListItem).text = btnTxt
 
-        holder.view.btnListItem.onClick {
-            //cnt.toast(btnTxt)
-            /*val i: Intent = Intent(cnt, BusViewMapActivity::class.java)
-            i.putExtra("BUS",btnTxt)
-            cnt.startActivity(i)*/
-            cnt.startActivity<BusViewMapActivity>("BUS" to btnTxt)
-
+        holder.view.findViewById<Button>(R.id.btnListItem).setOnClickListener {
+            Intent(cnt, BusViewMapActivity::class.java).apply {
+                putExtra("BUS", btnTxt)
+            }.also { cnt.startActivity(it) }
 
         }
 
