@@ -1,6 +1,6 @@
 package com.delzor.zb.csakbusz.adapters
 
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.View
 
 class CustomViewHolder(val view: View) : RecyclerView.ViewHolder(view)
