@@ -1,18 +1,17 @@
 package com.delzor.zb.csakbusz.adapters
 
 import android.graphics.Color
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.util.TypedValue
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.ImageView
+import android.widget.TextView
 import com.delzor.zb.csakbusz.Data
 import com.delzor.zb.csakbusz.Line
 import com.delzor.zb.csakbusz.R
 import com.delzor.zb.csakbusz.Utils
-import kotlinx.android.synthetic.main.list_item_sublines.view.*
-import org.jetbrains.anko.backgroundColor
-import org.jetbrains.anko.sdk25.coroutines.onClick
 
 class LineListAdapter(var itemList: MutableList<Line>, val onItmClick: (Line) -> Unit) : RecyclerView.Adapter<CustomViewHolder>() {
     var prev: View? = null
@@ -30,11 +29,12 @@ class LineListAdapter(var itemList: MutableList<Line>, val onItmClick: (Line) ->
         val cnt = holder.view.context
         val line = itemList[position]
 
-        holder.view.tvSublineName.text = line.name.replace("-", " - ")
-        holder.view.ivSublineIcon.setImageBitmap(Utils.circularBG(cnt, Color.parseColor(Data.KOLORS(position))))
 
-        holder.view.onClick {
-            holder.view.backgroundColor = Color.parseColor(Data.KOLORS(position))
+        holder.view.findViewById<TextView>(R.id.tvSublineName).text = line.name.replace("-", " - ")
+        holder.view.findViewById<ImageView>(R.id.ivSublineIcon).setImageBitmap(Utils.circularBG(cnt, Color.parseColor(Data.KOLORS(position))))
+
+        holder.view.setOnClickListener {
+            holder.view.setBackgroundColor(Color.parseColor(Data.KOLORS(position)))
             //Color.parseColor("#BBFF61")
             if (prev != null && prev != holder.view) {
                 val outValue = TypedValue()
