@@ -1,11 +1,11 @@
 package com.delzor.zb.csakbusz.adapters
 
 import android.graphics.Color
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.ViewGroup
+import android.widget.TextView
 import com.delzor.zb.csakbusz.R
-import kotlinx.android.synthetic.main.list_item_text.view.*
 
 class SimpleListAdapter(var itemList: MutableList<String>, var textSize: Float = 14f, var style: Boolean = false) : RecyclerView.Adapter<CustomViewHolder>() {
     override fun getItemCount(): Int {
@@ -22,13 +22,13 @@ class SimpleListAdapter(var itemList: MutableList<String>, var textSize: Float =
         val cnt = holder.view.context
         val text = itemList[position]
 
-        if (style && position % 2 != 0) {
-            holder.view.setBackgroundColor(Color.parseColor("#E1E5E2"))
+        holder.view.run{
+            if (style && position % 2 != 0)
+                setBackgroundColor(Color.parseColor("#E1E5E2"))
+            findViewById<TextView>(R.id.text1).also{
+                it.textSize = textSize
+                it.text = text
+            }
         }
-
-        holder.view.text1.textSize = textSize
-        holder.view.text1.text = text
-
-
     }
 }
